+++ packages/windie-sdk-js/src/transport/ManagedWindieAgentSession.ts
@@ -3,3 +3,13 @@
  */
 
 export * from './ManagedAgentSession.js';
+
+export {
+  ManagedAgentSession as ManagedWindieAgentSession,
+  createManagedAgentSession as createManagedWindieAgentSession,
+} from './ManagedAgentSession.js';
+
+export type {
+  ManagedAgentBackendEndpoint as WindieManagedBackendEndpoint,
+  ManagedAgentSessionOptions as ManagedWindieAgentSessionOptions,
+} from './ManagedAgentSession.js';